+++ jaxrs/src/main/java/de/taimos/dvalin/jaxrs/MapperFactory.java
@@ -25,6 +25,8 @@ import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.MapperFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
+import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
+import com.fasterxml.jackson.dataformat.yaml.YAMLGenerator;
 import com.fasterxml.jackson.datatype.guava.GuavaModule;
 import com.fasterxml.jackson.datatype.joda.JodaModule;
 
@@ -36,6 +38,23 @@ public final class MapperFactory {
 
     public static ObjectMapper createDefault() {
         ObjectMapper m = new ObjectMapper();
+        m.registerModule(new JodaModule());
+        m.registerModule(new GuavaModule());
+        m.setSerializationInclusion(Include.NON_NULL);
+        m.disable(SerializationFeature.FAIL_ON_EMPTY_BEANS);
+        m.disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+        m.enable(DeserializationFeature.USE_BIG_DECIMAL_FOR_FLOATS);
+        m.enable(MapperFeature.AUTO_DETECT_GETTERS);
+        return m;
+    }
+
+    public static ObjectMapper createDefaultYaml() {
+        YAMLFactory factory = new YAMLFactory();
+        factory.disable(YAMLGenerator.Feature.WRITE_DOC_START_MARKER);
+        factory.enable(YAMLGenerator.Feature.MINIMIZE_QUOTES);
+        factory.enable(YAMLGenerator.Feature.SPLIT_LINES);
+        factory.enable(YAMLGenerator.Feature.ALWAYS_QUOTE_NUMBERS_AS_STRINGS);
+        ObjectMapper m = new ObjectMapper(factory);
         m.registerModule(new JodaModule());
         m.registerModule(new GuavaModule());
         m.setSerializationInclusion(Include.NON_NULL);